google-cloud-embedding-gecko/requirements.txt
google-cloud-aiplatform
mteb
